Standardize library helper naming
